Define a specification for the expanding of nodes
Currently there is a pre-existing expansion library contained within a
separate project. It does not however a associated specification file.

Instead, the specification has been defined now before the library is
merged in. This way it can act as acceptance tests for the current state
of the library.

Then any future feature can be tested in a standard BDD style