Preform basic alphanumeric node name checking
The node must have an alphanumeric node include `_`. It is sensible to
enforce a node must be named sensible as special characters could affect
the parsing. I am not sure if `_` should be allowed at this stage, but
it won't affect the parser

NOTE: This doesn't completely fix the spec